backend/routing-main/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
requests
geojson
python-dotenv
shapely
gunicorn
